Add files via upload
Added Music, changed some minor stuff, added the "complete" level one and a majority of the dungeon assets. Added pixel healthbar and death screen that leads back to level select and main menu. Bug 1: After the enemy Knight attacks, even if the player is far outside of attack range, the enemy will continue to attack and deal damage until fatal to the player. Bug 2: when coming face to face with the frog enemy, if the player is actively moving when the projectile hits, the player takes 10 damage, but if the player is standing still the player dies in 2 shots, with the first taking a majority of player health.